Update README.md
Add clarifications that once .env is modified, the miner defaults to that one and the other command is only if you wish to mind to a different wallet than the one specified in the .env file.

I saw a lot of confusion about this on telegram. please confirm that my perceptions are correct.